[refactor] - 미팅룸 리스트 파일구조 변경
